iot_site/requirements.txt
django
django-bootstrap4
djangorestframework
pyyaml
gunicorn
